requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==43.0.1
Django==5.2.6
djangorestframework==3.15.2
idna==3.10
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.23
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.5.0
wheel==0.45.1
